backend/news-import/requirements.txt
psycopg2-binary>=2.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
